Level 1 Triggers working
the red, green, blue and purple triggers all have prefabs with their boxes including their scripts. Two new scripts added one for sector A door to check that the triggers in that section are all true and another script for sector B to open the final portal door if every trigger is set to true. Level 2 will still need the trigger implementation and new script for sector A and B door as its not the same as level 1.