level 1 and level 2 triggers complete
All the triggers in level 1 and level 2 work as intended there was an issue with array as it accepts other collision objects which messed with activating the triggers so made a check to check the tags of the boxes of the specific trigger which worked fine. There are new two scripts for the sector doors.