changes made to stairs based on feedback and made some tests on lasers
Made adjustments on the stairs in both levels to make it easier for players to go up and down on. carried out two laser tests to check the laser shoots laser by checking its position and to check if it hits a wall by checking checking it against the expected name.