Completed level 2 design and main menu
level 2 has been fully re-designed and main menu works as intended where play can change volume, quality, resolution or if the game is in full screen mode or not. Made a small fix which freezes time when player dies and when returned to main menu to and start game to unfreeze it. another minor change in gameOver and gameComplete where player can return to main menu.