CutScene animation
created all cutscenes for level 1 and level 2 including creating a class for triggers due to not all trigger have a cutscene so the ones which do use the cutSceneDoorTrigger. Updated any class which would use a cutScene such as sector doors or bridge triggers. Minor change to the main menu to allow you to exit game when quit button is clicked.